+++ src/main/java/gt/com/edu/security/dao/IUsuarioDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gt.com.edu.security.entity.Usuario;
 
-public interface IUsuarioDao extends JpaRepository<Usuario, Integer>{
+public interface IUsuarioDao extends JpaRepository<Usuario, Long>{
 
 }
